# apps/api/src/assets/openapi.yaml
openapi: 3.1.0
info:
  title: Performa AI API
  description: Unified AI API platform with pay-per-use billing
  version: 1.0.0
  contact:
    name: Performa AI
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /auth/keys/regenerate:
    post:
      tags: [Auth]
      summary: Regenerate API key
      operationId: regenerateApiKey
      security:
        - BearerAuth: []
      responses:
        "200":
          description: New API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey:
                    type: string
                    example: sk_live_abc123def456
        "401":
          description: Unauthorized
...
  /v1/knowledge/bases:
    post:
      tags: [Knowledge]
      summary: Create a knowledge base
      operationId: createKnowledgeBase
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "201":
          description: Knowledge base created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
    get:
      tags: [Knowledge]
      summary: List knowledge bases
      operationId: listKnowledgeBases
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: List of knowledge bases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeBase"
        "401":
          description: Unauthorized
  /v1/knowledge/bases/{id}:
    get:
      tags: [Knowledge]
      summary: Get a knowledge base
      operationId: getKnowledgeBase
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Knowledge base details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "401":
          description: Unauthorized
        "404":
          description: Knowledge base not found
    put:
      tags: [Knowledge]
      summary: Update a knowledge base
      operationId: updateKnowledgeBase
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKnowledgeBaseRequest"
      responses:
        "200":
          description: Knowledge base updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Knowledge base not found
    delete:
      tags: [Knowledge]
      summary: Delete a knowledge base
      operationId: deleteKnowledgeBase
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Knowledge base deleted
        "401":
          description: Unauthorized
        "404":
          description: Knowledge base not found
  /v1/knowledge/bases/{id}/upload:
    post:
      tags: [Knowledge]
      summary: Upload a markdown file to a knowledge base
      description: Uploads a .md file, chunks it, generates embeddings, and stores in the knowledge base. Rate limited to 5 requests per minute.
      operationId: uploadMarkdown
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Markdown file (.md, max 10MB)
      responses:
        "200":
          description: File processed and chunks created
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunksCreated:
                    type: integer
                  s3Key:
                    type: string
        "400":
          description: Invalid file (not .md or exceeds size limit)
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded
  /v1/knowledge/bases/{id}/chunks:
    post:
      tags: [Knowledge]
      summary: Add text chunks to a knowledge base
      description: Manually add text chunks with optional metadata. Rate limited to 10 requests per minute.
      operationId: addChunks
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChunksRequest"
      responses:
        "201":
          description: Chunks added
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunksCreated:
                    type: integer
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "429":
          description: Rate limit exceeded
    get:
      tags: [Knowledge]
      summary: List chunks in a knowledge base
      operationId: listChunks
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkList"
        "401":
          description: Unauthorized
  /v1/knowledge/bases/{id}/chunks/{chunkId}:
    delete:
      tags: [Knowledge]
      summary: Delete a chunk
      operationId: deleteChunk
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: chunkId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Chunk deleted
        "401":
          description: Unauthorized
        "404":
          description: Chunk not found
  /v1/knowledge/search:
    post:
      tags: [Knowledge]
      summary: Search knowledge base chunks by semantic similarity
      operationId: searchKnowledgeChunks
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchChunksRequest"
      responses:
        "200":
          description: Search results ranked by similarity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API key with Bearer prefix: `Bearer sk_live_*`"
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: mypassword123
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 1
          example: mypassword123
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        apiKey:
          type: string
          description: API key for chat completions (only on register)
          example: sk_live_abc123def456
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        balance:
          type: string
          description: Current balance in USD
        apiKey:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ValidationError:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
...
    KnowledgeBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
    UpdateKnowledgeBaseRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        active:
          type: boolean
    AddChunksRequest:
      type: object
      required: [chunks]
      properties:
        chunks:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [content]
            properties:
              content:
                type: string
                minLength: 1
                maxLength: 100000
              metadata:
                type: object
                additionalProperties: true
    SearchChunksRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 10000
        knowledgeBaseId:
          type: string
          format: uuid
          description: Optional filter to search within a specific knowledge base
        topK:
          type: integer
          minimum: 1
          maximum: 50
          default: 5
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        metadata:
          type: object
          nullable: true
        similarity:
          type: number
          description: Cosine similarity score (0-1)
        knowledgeBaseId:
          type: string
          format: uuid
    ChunkList:
      type: object
      properties:
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeChunk"
        total:
          type: integer
    KnowledgeChunk:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        metadata:
          type: object
          nullable: true
        tokenCount:
          type: integer
        createdAt:
          type: string
          format: date-time